Move files to core directory
